Rename backend controller and model files
